Rename project to Wyoming Apple Speech; add TTS-era README
The server now speaks as well as listens, so "Apple STT" no longer covers
it. Renames the Python package, Homebrew formulas (+beta), launchd label,
config/log paths, and all URLs to wyoming-apple-speech. The apple-stt and
apple-tts binaries keep their per-function names. README is restructured
around a one-line install, HomePod-class positioning, examples, and a
troubleshooting table.
