+++ .github/workflows/release.yml
@@ -23,16 +23,16 @@ jobs:
           VERSION="${TAG#v}"
 
           # A -beta suffix (e.g. v1.2.0-beta.1) selects the beta channel: a GitHub
-          # prerelease and a separate wyoming-apple-stt-beta formula that coexists with,
+          # prerelease and a separate wyoming-apple-speech-beta formula that coexists with,
           # and conflicts with, the stable one.
           if [[ "${TAG}" == *-beta* ]]; then
             CHANNEL="beta"
-            FORMULA_NAME="wyoming-apple-stt-beta"
-            FORMULA_CLASS="WyomingAppleSttBeta"
+            FORMULA_NAME="wyoming-apple-speech-beta"
+            FORMULA_CLASS="WyomingAppleSpeechBeta"
           else
             CHANNEL="stable"
-            FORMULA_NAME="wyoming-apple-stt"
-            FORMULA_CLASS="WyomingAppleStt"
+            FORMULA_NAME="wyoming-apple-speech"
+            FORMULA_CLASS="WyomingAppleSpeech"
           fi
 
           echo "tag=${TAG}" >> "$GITHUB_OUTPUT"
@@ -49,7 +49,7 @@ jobs:
         id: tarball
         run: |
           ./packaging/build-release-tarball.sh "${{ steps.version.outputs.version }}"
-          TARBALL="dist/wyoming-apple-stt-${{ steps.version.outputs.version }}.tar.gz"
+          TARBALL="dist/wyoming-apple-speech-${{ steps.version.outputs.version }}.tar.gz"
           SHA="$(shasum -a 256 "${TARBALL}" | awk '{print $1}')"
           echo "path=${TARBALL}" >> "$GITHUB_OUTPUT"
           echo "sha256=${SHA}" >> "$GITHUB_OUTPUT"
@@ -81,7 +81,7 @@ jobs:
         run: |
           set -euo pipefail
 
-          URL="https://github.com/FI-153/wyoming-apple-stt/releases/download/${TAG}/wyoming-apple-stt-${VERSION}.tar.gz"
+          URL="https://github.com/FI-153/wyoming-apple-speech/releases/download/${TAG}/wyoming-apple-speech-${VERSION}.tar.gz"
 
           URL="${URL}" python3 - <<'PY'
           import os
@@ -91,7 +91,7 @@ jobs:
           # binaries and launchd service. conflicts_with is only enforced from the declaring
           # side, so each formula names the other.
           because = "both install the same binaries and launchd service"
-          other = "wyoming-apple-stt" if os.environ["CHANNEL"] == "beta" else "wyoming-apple-stt-beta"
+          other = "wyoming-apple-speech" if os.environ["CHANNEL"] == "beta" else "wyoming-apple-speech-beta"
           conflicts = f'conflicts_with "{other}", because: "{because}"'
 
           tpl = Path("packaging/formula.rb.template").read_text()
