+++ com.wyoming-apple-speech.plist.template
@@ -4,13 +4,13 @@
 <plist version="1.0">
 <dict>
     <key>Label</key>
-    <string>com.wyoming-apple-stt</string>
+    <string>com.wyoming-apple-speech</string>
 
     <key>ProgramArguments</key>
     <array>
         <string>__VENV_PYTHON__</string>
         <string>-m</string>
-        <string>wyoming_apple_stt</string>
+        <string>wyoming_apple_speech</string>
         <string>--uri</string>
         <string>tcp://0.0.0.0:__PORT__</string>
         <string>--apple-stt-bin</string>
