+++ wyoming_apple_speech/__init__.py
@@ -1,0 +1,1 @@
+"""Wyoming Apple Speech — on-device speech recognition and Siri TTS for Home Assistant."""
