--- wyoming_apple_stt/__init__.py
@@ -1,1 +1,0 @@
-"""Wyoming Apple STT — macOS on-device speech recognition for Home Assistant."""
